requirements.txt
# Core libraries
langchain==0.1.4
sentence-transformers==2.5.1  # Free local embeddings
chromadb==0.4.22  # Free vector database
pydantic==2.5.2

# Data processing
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
youtube-transcript-api==0.6.1

# LLM (choose based on your hardware)
ctransformers==0.2.27  # Run smaller LLMs locally
llama-cpp-python==0.2.27  # Optional for Llama models

# Frontend
streamlit==1.28.2

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1